append-script: fmt

+++ mpv/playlist_ctl/append_video.sh
@@ -5,15 +5,16 @@ MPV_SOCKET_FILE=/tmp/mpv.sock
 
 notify() { [ -n "$1" ] && notify-send -i mpv -a append-script "$1"; }
 
-[ -f "$MPV_START_LOCK" ] && {
-    notify 'MPV process already running...'
+if [ -f "$MPV_START_LOCK" ]; then
+    notify 'mpv process already running...'
     exit 1
-}
+fi
 
 cleanup() { [ -f "$MPV_START_LOCK" ] && rm "$MPV_START_LOCK"; }
 trap cleanup EXIT
 
-echo '{"command":["get_property","pause"]}' | nc -NU "$MPV_SOCKET_FILE" >/dev/null 2>&1 || {
+if ! echo '{"command":["get_property","pause"]}' | nc -NU "$MPV_SOCKET_FILE" >/dev/null 2>&1; then
+    notify 'starting mpv...'
     touch "$MPV_START_LOCK"
     [ -S "$MPV_SOCKET_FILE" ] && rm "$MPV_SOCKET_FILE"
     setsid -f mpv --idle --no-terminal --input-ipc-server="$MPV_SOCKET_FILE" >/dev/null 2>&1
@@ -27,12 +28,12 @@ echo '{"command":["get_property","pause"]}' | nc -NU "$MPV_SOCKET_FILE" >/dev/nu
                 nc -NU "$MPV_SOCKET_FILE" >/dev/null 2>&1 && break
         fi
     done
-}
+fi
 
-[ -S "$MPV_SOCKET_FILE" ] || {
+if [ ! -S "$MPV_SOCKET_FILE" ]; then
     notify "can't find $MPV_SOCKET_FILE"
     killall mpv
     exit 1
-}
+fi
 
 notify "$(playlist-ctl -a "$1" 2>&1)"
